+++ spring-cloud-tencent-commons/src/main/java/com/tencent/cloud/common/tsf/ConditionalOnTsfEnabled.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.tencent.cloud.polaris.context.tsf;
+package com.tencent.cloud.common.tsf;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
