+++ spring-cloud-tencent-commons/src/main/java/com/tencent/cloud/common/util/inet/PolarisInetUtilsAutoConfiguration.java
@@ -17,6 +17,8 @@
 
 package com.tencent.cloud.common.util.inet;
 
+import com.tencent.cloud.common.tsf.ConditionalOnTsfEnabled;
+
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.cloud.commons.util.InetUtilsProperties;
 import org.springframework.context.annotation.Bean;
@@ -29,6 +31,7 @@ import org.springframework.context.annotation.Configuration;
  * @author Haotian Zhang
  */
 @Configuration(proxyBeanMethods = false)
+@ConditionalOnTsfEnabled
 public class PolarisInetUtilsAutoConfiguration {
 
 	@Bean
